+++ lib/td_dd/data_structures/search/aggregations.ex
@@ -41,7 +41,10 @@ defmodule TdDd.DataStructures.Search.Aggregations do
     content
     |> Format.flatten_content_fields()
     |> Enum.flat_map(fn
-      %{"name" => field, "type" => type} when type in ["domain", "system"] ->
+      %{"name" => field, "type" => "domain"} ->
+        [{field, %{terms: %{field: "content.#{field}", size: 50}, meta: %{type: "domain"}}}]
+
+      %{"name" => field, "type" => "system"} ->
         [{field, nested_agg(field)}]
 
       %{"name" => field, "type" => "user"} ->